% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbs_smooth_data}
\alias{bbs_smooth_data}
\title{Pacific Wren bbs Smooth output from 1968 to 2022}
\format{
\subsection{\code{bbs_smooth_data}}{

A tibble with 55 rows and 4 columns:
\describe{
\item{year}{year of indices}
\item{index}{annual index}
\item{index_q_0.025}{confidence interval 2.5\%}
\item{index_q_0.975}{confidence interval 97.5\%}
...
}
}
}
\source{
\url{https://github.com/bbsBayes/bbsBayes2/tree/main}
}
\usage{
bbs_smooth_data
}
\description{
An example of posterior draws from smooth component of BBS model for Pacific Wren
}
\keyword{datasets}
