% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdtrends-package.R
\docType{package}
\name{birdtrends-package}
\alias{birdtrends}
\alias{birdtrends-package}
\title{birdtrends: Estimate Trends Based On Modelled Annual Indices Or Related Inputs}
\description{
This package provides a flexible framework to generate trends, based on actual or modeled annual indices data. This package was designed to undertake assessment of 2016 Partners In Flight Landbird Conservation objectives.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ninoxconsulting/birdtrends}
  \item Report bugs at \url{https://github.com/ninoxconsulting/birdtrends/issues}
}

}
\author{
\strong{Maintainer}: Genevieve Perkins \email{genevieve.c.perkins@gmail.com} (\href{https://orcid.org/0000-0003-2663-5904}{ORCID})

Authors:
\itemize{
  \item Adam C. Smith \email{adam.smith2@canada.ca}
}

}
\keyword{internal}
