% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probs.R
\name{calculate_probs}
\alias{calculate_probs}
\title{Calculate probability of meeting targets}
\usage{
calculate_probs(
  projected_trends = preds_sm,
  ref_year = 2014,
  targ_year = 2026,
  prob_decrease = NULL,
  prob_increase = NULL
)
}
\arguments{
\item{projected_trends}{tibble with all draws, years, projected trends, output of proj_trend function}

\item{ref_year}{numeric year in which change is measured from i.e. 2014}

\item{targ_year}{numeric year in which target will be reached i.e. 2026}

\item{prob_decrease}{numeric vector of predicted target increase. i.e c(25, 35). Determine the probability that trend will increase 25\% to 35\% by target year}

\item{prob_increase}{numeric vector of predicted target decrease. i.e c(25, 35). Determine the probability that trend will decline by 25\% to 35\% by target year}
}
\value{
tibble with the average percent change in addition to probability of increase or decrease
}
\description{
Calculate probability of meeting targets
}
\examples{
\dontrun{
  test_st <- calculate_probs(predicted_trends = preds_sm,
        ref_year = 2014,
        targ_year = 2026,
        prob_increase = c(targ$st_pop_pc_lower,targ$st_pop_pc_uppper))
}
}
