% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{fit_gam}
\alias{fit_gam}
\title{Fit gam models on posterior draws}
\usage{
fit_gam(indata, start_yr = NA, end_yr = NA, n_knots = 5)
}
\arguments{
\item{indata}{dataframe of posterior draws. Each column is a year, each row is a posterior draw}

\item{start_yr}{numeric year at which to start model. Default is the first year available}

\item{end_yr}{numeric year at which to end model. Default is the last year available}

\item{n_knots}{number of knots used in the gam model, default is 5}
}
\value{
dataframe with predicted
}
\description{
Fit gam models on posterior draws
}
\examples{
\dontrun{
pred_dataset <- fit_gam(indat2, start_yr = 1990, end_yr = 2020, n_knots = 14)
}
}
