% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hgam.R
\name{fit_hgam}
\alias{fit_hgam}
\title{Fit hierarchical GAM using annual indices and confidence interval}
\usage{
fit_hgam(indata, start_yr = NA, end_yr = NA, n_knots = NA, longform = TRUE)
}
\arguments{
\item{indata}{dataframe with annual indices by year (row) and confidence intervals}

\item{start_yr}{numeric year at which to start model. Default is the first year available}

\item{end_yr}{numeric year at which to end model. Default is the last year available}

\item{n_knots}{number of knots used in the HGAM model, using default of one knot per 4 years of data}

\item{longform}{TRUE/FALSE the output will be converted to a longform tibble with columns draw, year, proj_y. Default = TRUE}
}
\value{
tibble with modeled HGAM indices for given years
}
\description{
Fit hierarchical GAM using annual indices and confidence interval
}
\examples{
\dontrun{
outsmooth <- fit_hgam(indata = indat1, start_yr = 1970, end_yr = 2020, n_knots = 5)
}
}
