% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_smooths.R
\name{fit_smooths}
\alias{fit_smooths}
\title{Extract Smooths from BBS model run}
\usage{
fit_smooths(indata, start_yr = NA, end_yr = NA)
}
\arguments{
\item{indata}{dataframe of generated smooths. Each column is a year, each row is a predicted smooth}

\item{start_yr}{numeric year at which to start, if subset is required. Default is the last year available}

\item{end_yr}{numeric year at which to end, if subset is required. Default is the last year available}
}
\value{
dataframe with smooths generated for each year of the selected range
}
\description{
Extract Smooths from BBS model run
}
\examples{
\dontrun{
smooths <- fit_smooths(indat3, start_yr = 1990, end_yr = 2020)
}
}
