% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_targets.R
\name{get_targets}
\alias{get_targets}
\title{Get target values for plotting}
\usage{
get_targets(
  model_indices = ldf,
  ref_year = 2014,
  st_year = 2026,
  st_lu_target_pc = -2,
  st_up_target_pc = 1,
  lt_year = NA,
  lt_lu_target_pc = NA,
  lt_up_target_pc = NA
)
}
\arguments{
\item{model_indices}{tibble of estimated indices drawn from fit_* function. Columns are draw, year, and proj_y}

\item{ref_year}{numeric year in which targets are compared to, default = 2014}

\item{st_year}{numeric short term year for which targets will be estimated}

\item{st_lu_target_pc}{numeric lower confidence interval of percentage of short term population change from reference year to st_year}

\item{st_up_target_pc}{numeric upper confidence interval of percentage of short term population change from reference year to st_year}

\item{lt_year}{numeric long term year for which targets will be estimated}

\item{lt_lu_target_pc}{numeric lower confidence interval of percentage of long term population change from reference year to lt_year}

\item{lt_up_target_pc}{numeric upper confidence interval of percentage of long term population change from reference year to lt_year}
}
\value{
tibble with upper and lower target values
}
\description{
Get target values for plotting
}
\examples{
\dontrun{
  hgams_plot <- get_targets(model_indices = ldf, ref_year = 2014,
  st_year = 2026, st_lu_target_pc = -2,st_up_target_pc = 1,
   lt_year = 2046,  lt_lu_target_pc = 5,lt_up_target_pc = 15)
}
}
