% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_hgam.R
\name{prep_hgam}
\alias{prep_hgam}
\title{Prepare GAM components for running stan model}
\usage{
prep_hgam(
  orig.preds = NA,
  nknots = 6,
  predpoints = NULL,
  npredpoints = 100,
  sm_name = "year"
)
}
\arguments{
\item{orig.preds}{vector of years to be modeled}

\item{nknots}{numeric value for number of knots in model}

\item{predpoints}{values to be predicted. Default is NULL}

\item{npredpoints}{number of values to be predicted. Default is 100}

\item{sm_name}{name of model column outputs, default is "year}
}
\value{
a list containing components of  GAM model to be run within STAN
}
\description{
Prepare GAM components for running stan model
}
\examples{
\dontrun{
aa <- prep_hgam(out$year, nknots = 6, predpoints = NULL, npredpoints = 100, sm_name = "year")
}
}
