% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_change.R
\name{trend_change}
\alias{trend_change}
\title{Calculate change in trend between years}
\usage{
trend_change(projected_trends = preds_sm, ref_year = 2014, targ_year = 2026)
}
\arguments{
\item{projected_trends}{tibble with all draws, years, projected trends, output of proj_trend function}

\item{ref_year}{numeric year in which change is measured from i.e. 2014}

\item{targ_year}{numeric year in which target will be reached i.e. 2026}
}
\value{
tibble with the average percent change in addition to probability of increase or decrease
}
\description{
Calculate change in trend between years
}
\examples{
\dontrun{
  test_st <- trend_change(predicted_trends = preds_sm,
        ref_year = 2014,
        targ_year = 2026)
}
}
